+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 ### Improvements
 - Removed the PvP balance-loss feature and moved it to the new bounty add-on.
 - `/orders claim` has been changed to `/deliveries`.
+- Added filters to the transactions UI.
 
 ### Fixes